+++ pyproject.toml
@@ -114,10 +114,6 @@ show_error_codes = true
 pretty = true
 plugins = ["pydantic.mypy"]
 
-[tool.pydantic-mypy]
-init_forbid_extra = true
-init_typed = false
-warn_required_dynamic_aliases = true
 
 # https://docs.pytest.org/
 [tool.pytest.ini_options]
