--- src/ndv/model/__init__.py
@@ -1,1 +1,0 @@
-"""Models for NDV."""
