+++ src/ndv/models/_base_model.py
@@ -1,0 +1,14 @@
+from typing import ClassVar
+
+from psygnal import SignalGroupDescriptor
+from pydantic import BaseModel, ConfigDict
+
+
+class NDVModel(BaseModel):
+    """Base eventd model for NDV models."""
+
+    model_config = ConfigDict(
+        validate_assignment=True,
+        validate_default=True,
+    )
+    events: ClassVar[SignalGroupDescriptor] = SignalGroupDescriptor()
